% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rent_or_buy.R
\name{rent_or_buy}
\alias{rent_or_buy}
\title{rent_or_buy function
Runs the simulation on whats better: renting or buying}
\usage{
rent_or_buy(price = 750000, down = price * 0.2, interest = 0.06,
  loan_yrs = 30, tax = 0.0125, maintain = price * 0.005, hoa = 0,
  insur = 800, apprec = 0.03, marg_tax = 0.3, inflation = 0.02,
  rent = 2500, rent_inflation = 0.03, cash_return = 0.12)
}
\arguments{
\item{price, }{price of the house}

\item{down, }{size of sample for picking the tickers}

\item{interest, }{size of sample for picking the tickers}

\item{loan_yrs, }{size of sample for picking the tickers}

\item{tax, }{size of sample for picking the tickers}

\item{maintain, }{size of sample for picking the tickers}

\item{hoa, }{size of sample for picking the tickers}

\item{insur, }{size of sample for picking the tickers}

\item{apprec, }{size of sample for picking the tickers}

\item{marg_tax, }{size of sample for picking the tickers}

\item{inflation, }{size of sample for picking the tickers}

\item{rent, }{size of sample for picking the tickers}

\item{rent_inflation, }{size of sample for picking the tickers}

\item{cash_return, }{size of sample for picking the tickers}
}
\description{
rent_or_buy function
Runs the simulation on whats better: renting or buying
}
\examples{
rent_or_buy()
}
\keyword{arith}

