% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunExample.R
\name{runExample}
\alias{runExample}
\title{Example of rent_or_buy computation
Runs the simulation on whats better: renting or buying}
\usage{
runExample()
}
\description{
Example of rent_or_buy computation
Runs the simulation on whats better: renting or buying
}
\keyword{arith}

