% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock_pick.R
\name{stock_pick}
\alias{stock_pick}
\title{The stock_pick function
Constructs the optimal portofolio once the stock tickers are provided or designs one based on sample of tickers.}
\usage{
stock_pick(stocks = NULL, stocks_lim = 15, freq = "week",
  start = "2010-01-01", min_pos = 4)
}
\arguments{
\item{stocks, }{stock tickers for portofolio, can be null}

\item{stocks_lim, }{size of sample for picking the tickers}

\item{freq, }{size of sample for picking the tickers}

\item{start, }{size of sample for picking the tickers}

\item{min_pos, }{size of sample for picking the tickers}
}
\description{
The stock_pick function
Constructs the optimal portofolio once the stock tickers are provided or designs one based on sample of tickers.
}
\examples{
stock_pick()

}
\keyword{arith}

