% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stocks.R
\docType{data}
\name{stocks}
\alias{stocks}
\title{Sample of stock tickers.}
\format{A character vector:
\describe{
  \item{name}{price, in US dollars}
  ...
}}
\source{
\url{http://www.diamondse.info/}
}
\usage{
stocks
}
\description{
A dataset containing the prices and other attributes of almost 54,000
diamonds.
}
\keyword{datasets}

